Changed decimals to floats. Hopefully this will fix the Sequelize/Pos…

+++ backend/db/migrations/20230811041709-create-event.js
@@ -42,7 +42,7 @@ module.exports = {
         type: Sequelize.INTEGER,
       },
       price: {
-        type: Sequelize.DECIMAL(8, 2),
+        type: Sequelize.FLOAT,
       },
       startDate: {
         type: Sequelize.DATE,
